Fix/Reorganize network unit tests
