+++ apps/network/tests/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+import os
+from grid_network import create_app
+
+
+@pytest.fixture(scope="function", autouse=True)
+def app():
+    db_path = "sqlite:///:memory:"
+    return create_app(debug=True, db_config={"SQLALCHEMY_DATABASE_URI": db_path})
+
+@pytest.fixture
+def client(app):
+    return app.test_client()
